modularizing the classes into manageable packages
