--- src/main/java/com/creativityskills/jotech/bean/ProductBeanI.java
@@ -1,7 +1,0 @@
-package com.creativityskills.jotech.bean;
-
-
-import com.creativityskills.jotech.model.Product;
-
-public interface ProductBeanI extends BeanI<Product> {
-}
