+++ src/main/java/com/creativityskills/jotech/bean/machine/VendingMachineI.java
@@ -1,4 +1,4 @@
-package com.creativityskills.jotech.bean;
+package com.creativityskills.jotech.bean.machine;
 
 import com.creativityskills.jotech.model.Denomination;
 import com.creativityskills.jotech.model.Product;
@@ -11,6 +11,6 @@ import java.util.Map;
 public interface VendingMachineI {
     BigDecimal calculateRequiredAmount(Product product, int quantity);
 
-    boolean makeSale(Product product, int quantity, Map<Denomination,Integer> denominations) throws Exception;
+    boolean makeSale(Product product, int quantity, Map<Denomination, Integer> denominations) throws Exception;
 
 }
