+++ src/main/java/com/creativityskills/jotech/bean/product/ProductBeanI.java
@@ -1,0 +1,14 @@
+package com.creativityskills.jotech.bean.product;
+
+
+import com.creativityskills.jotech.bean.crud.BeanI;
+import com.creativityskills.jotech.model.Product;
+
+import javax.ejb.Local;
+import java.util.List;
+
+@Local
+public interface ProductBeanI extends BeanI<Product> {
+
+    List<Product> getProducts();
+}
