+++ src/main/java/com/creativityskills/jotech/bean/util/MoneyConvertorI.java
@@ -1,4 +1,4 @@
-package com.creativityskills.jotech.bean;
+package com.creativityskills.jotech.bean.util;
 
 import com.creativityskills.jotech.model.Denomination;
