+++ src/main/java/com/creativityskills/jotech/model/CashDrawer.java
@@ -8,6 +8,7 @@ public class CashDrawer {
     public int getId() {
         return id;
     }
+
     public void setId(int id) {
         this.id = id;
     }
