+++ src/main/java/com/creativityskills/jotech/model/Stock.java
@@ -1,8 +1,5 @@
 package com.creativityskills.jotech.model;
 
-import javax.inject.Named;
-
-
 public class Stock {
     private int id;
     private Product product;
